Refactor agent governance validator into modular classes
